Implemented locale agnostic fix for atlas element bounds display
The original implementation for rendering bounding boxes on each asset
using an atlas was actually only broken for users in some non-European
(namely US) locales.

System.Convert.ToDouble(String) is localized and defaults to using the
current thread's "culture" (locale) which itself is defaulted based on
the host system at runtime. So, in Matt's last revision:

double u1 = Convert.ToDouble(xChild.Attributes.GetNamedItem("u1")
                                              .Value.Replace(".",","));

was a problem because it was hard-coded to only work for European-like
locales. On his system the decimal points would be swapped for commas
and then Convert.ToDouble(String) would correctly parse the float
string since his locale uses commas as the fractional delimiter.

What this means though, is that while removing the Replace(".",",")
statements causes ReadAtlasData to parse the atlas files correctly
for US style locales, it also breaks the parsing for the comma
based ones.

The solution is to use System.Convert.ToDouble(String, IFormatProvider)
which lets you specify a locale to use via anything that implements
the IFormatProvider interface.

This change facilitates said fix by providing a static instance of
CultureInfo("en-US") "GAME_LOCALE" under Util in the KleiLib library , which can be used to force
System functions that take locale into account to respect the game's
locale.